Refactor out `lovatAPI.getScouterOverviews()`
