+++ lib/reusable/lovat_api/lovat_api.dart
@@ -133,31 +133,6 @@ class LovatAPI {
 
   // MARK: Endpoints
 
-  Future<List<ScouterOverview>> getScouterOverviews() async {
-    final tournament = await Tournament.getCurrent();
-
-    final response = await lovatAPI.get(
-      '/v1/manager/scouterspage',
-      query: {
-        if (tournament != null) 'tournamentKey': tournament.key,
-      },
-    );
-
-    if (response?.statusCode != 200) {
-      try {
-        throw LovatAPIException(jsonDecode(response!.body)['displayError']);
-      } on LovatAPIException {
-        rethrow;
-      } catch (_) {
-        throw Exception('Failed to get scouter overviews');
-      }
-    }
-
-    final json = jsonDecode(response!.body) as List<dynamic>;
-
-    return json.map((e) => ScouterOverview.fromJson(e)).toList();
-  }
-
   Future<void> addScouter(String name) async {
     final response = await lovatAPI.post(
       '/v1/manager/scouterdashboard',
@@ -382,29 +357,6 @@ class ScouterPageMinimalScoutReportInfo {
       ),
       reportId: json['uuid'],
       teamNumber: json['teamMatchData']['teamNumber'],
-    );
-  }
-}
-
-class ScouterOverview {
-  const ScouterOverview({
-    required this.totalMatches,
-    required this.missedMatches,
-    required this.scout,
-  });
-
-  final int totalMatches;
-  final int missedMatches;
-  final Scout scout;
-
-  factory ScouterOverview.fromJson(Map<String, dynamic> json) {
-    return ScouterOverview(
-      totalMatches: json['matchesScouted'],
-      missedMatches: json['missedMatches'] ?? 0,
-      scout: Scout(
-        id: json['scouterUuid'],
-        name: json['scouterName'],
-      ),
     );
   }
 }
